refactor: move task domain to features/task/
