+++ src/features/column/components/ColumnContainer.tsx
@@ -26,7 +26,7 @@ import {
 } from "@/shared/index";
 import { useKanban, type ColumnType, type Task } from "@/features/board/index";
 import { EditableColumnTitle } from "./EditableColumnTitle/EditableColumnTitle";
-import { CreateTaskSheet, TaskCard } from "@/index";
+import { CreateTaskSheet, TaskCard } from "@/features/task/index";
 
 interface Props {
   column: ColumnType;
